feat: Added instrumentation for chat completions create streams
