+++ lib/instrumentation/openai.js
@@ -14,6 +14,7 @@ const {
 const LlmTrackedIds = require('../../lib/llm-events/tracked-ids')
 
 const MIN_VERSION = '4.0.0'
+const MIN_STREAM_VERSION = '4.12.2'
 const { AI } = require('../../lib/metrics/names')
 const semver = require('semver')
 
@@ -135,18 +136,27 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
    * @param {TraceSegment} params.segment active segment from chat completion
    * @param {object} params.request chat completion params
    * @param {object} params.response chat completion response
-   * @param {boolean} [params.hasError=false] indicates that the completion
-   * resulted in some sort of error case
-   * @returns {LlmChatCompletionSummary} A summary object.
+   * @param {boolean} params.err err if it exists
    */
-  function recordChatCompletionMessages({ segment, request, response, hasError = false }) {
+  function recordChatCompletionMessages({ segment, request, response, err }) {
+    if (!response) {
+      // If we get an error, it is possible that `response = null`.
+      // In that case, we define it to be an empty object.
+      response = {}
+    }
+
+    response.headers = segment[openAiHeaders]
+    response.api_key = segment[openAiApiKey]
     const tx = segment.transaction
+    // explicitly end segment to consistent duration
+    // for both LLM events and the segment
+    segment.end()
     const completionSummary = new LlmChatCompletionSummary({
       agent,
       segment,
       request,
       response,
-      withError: hasError
+      withError: err != null
     })
 
     // Only take the first response message and append to input messages
@@ -168,7 +178,60 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
 
     recordEvent('LlmChatCompletionSummary', completionSummary)
 
-    return completionSummary
+    if (err) {
+      const llmError = new LlmErrorMessage({ cause: err, summary: completionSummary, response })
+      shim.agent.errors.add(segment.transaction, err, llmError)
+    }
+
+    delete response.headers
+    delete response.api_key
+  }
+
+  /*
+   * Chat completions create can return a stream once promise resolves
+   * This wraps the iterator which is a generator function
+   * We will call the original iterator, intercept chunks and yield
+   * to the original. On complete we will construct the new message object
+   * with what we have seen in the stream and create the chat completion
+   * messages
+   *
+   */
+  function instrumentStream({ request, response, segment }) {
+    shim.wrap(response, 'iterator', function wrapIterator(shim, orig) {
+      return async function* wrappedIterator() {
+        let content = ''
+        let role = ''
+        let chunk
+        let err
+        try {
+          const iterator = orig.apply(this, arguments)
+
+          for await (chunk of iterator) {
+            if (chunk.choices[0]?.delta?.role) {
+              role = chunk.choices[0]?.delta?.role
+            }
+
+            content += chunk.choices[0]?.delta?.content ?? ''
+            yield chunk
+          }
+        } catch (streamErr) {
+          err = streamErr
+          throw err
+        } finally {
+          chunk.choices[0].message = { role, content }
+          // update segment duration since we want to extend the time it took to
+          // handle the stream
+          segment.touch()
+
+          recordChatCompletionMessages({
+            segment,
+            request,
+            response: chunk,
+            err
+          })
+        }
+      }
+    })
   }
 
   /**
@@ -182,34 +245,28 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
     'create',
     function wrapCreate(shim, create, name, args) {
       const [request] = args
+      if (request.stream && semver.lt(shim.pkgVersion, MIN_STREAM_VERSION)) {
+        shim.logger.warn(
+          `Instrumenting chat completion streams is only supported with openai version ${MIN_STREAM_VERSION}+.`
+        )
+        return
+      }
+
       return {
         name: `${AI.OPEN_AI}/Chat/Completions/Create`,
         promise: true,
         // eslint-disable-next-line max-params
         after(_shim, _fn, _name, err, response, segment) {
-          if (!response) {
-            // If we get an error, it is possible that `response = null`.
-            // In that case, we define it to be an empty object.
-            response = {}
+          if (request.stream) {
+            instrumentStream({ request, response, segment })
+          } else {
+            recordChatCompletionMessages({
+              segment,
+              request,
+              response,
+              err
+            })
           }
-          response.headers = segment[openAiHeaders]
-          response.api_key = segment[openAiApiKey]
-
-          const summary = recordChatCompletionMessages({
-            segment,
-            request,
-            response,
-            hasError: err != null
-          })
-
-          if (err) {
-            const llmError = new LlmErrorMessage({ cause: err, summary, response })
-            shim.agent.errors.add(segment.transaction, err, llmError)
-          }
-
-          // cleanup keys on response before returning to user code
-          delete response.api_key
-          delete response.headers
         }
       }
     }
@@ -236,6 +293,9 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
           }
           response.headers = segment[openAiHeaders]
           response.api_key = segment[openAiApiKey]
+          // explicitly end segment to get consistent duration
+          // for both LLM events and the segment
+          segment.end()
           const embedding = new LlmEmbedding({
             agent,
             segment,
