+++ lib/shim/shim.js
@@ -30,7 +30,7 @@ const fnApply = Function.prototype.apply
  * @param {string}  moduleName    - The name of the module being instrumented.
  * @param {string}  resolvedName  - The full path to the loaded module.
  * @param {string}  shimName      - Used to persist shim ids across different instances. This is
- * @param {string}  pkgVersion    - version of module
+ * @param {string} pkgVersion     - version of package getting instrumented
  * applicable to instrument that compliments each other across libraries(i.e - koa + koa-route/koa-router)
  */
 function Shim(agent, moduleName, resolvedName, shimName, pkgVersion) {
