+++ test/versioned/openai/mock-server.js
@@ -9,7 +9,8 @@ module.exports = openaiMockServer
 
 const http = require('node:http')
 const { Readable } = require('node:stream')
-const RESPONSES = require('./openai-responses')
+const RESPONSES = require('./mock-responses')
+const crypto = require('crypto')
 
 /**
  * Build a mock server that listens on a 127.0.0.1 and a random port that
@@ -65,7 +66,7 @@ function handler(req, res) {
       return
     }
 
-    const { headers, code, body } = RESPONSES.get(prompt)
+    const { headers, code, body, streamData } = RESPONSES.get(prompt)
     for (const [key, value] of Object.entries(headers)) {
       res.setHeader(key, value)
     }
@@ -75,38 +76,55 @@ function handler(req, res) {
       // OpenAI streamed responses are double newline delimited lines that
       // are prefixed with the string `data: `. The end of the stream is
       // terminated with a `done: [DONE]` string.
-      const parts = body.split(' ')
-      let i = 0
-      const outStream = new Readable({
-        read() {
-          if (i < parts.length) {
-            const content = parts.length - 1 === i ? parts[i] : `${parts[i]} `
-            const chunk = JSON.stringify({
-              id: 'chatcmpl-8MzOfSMbLxEy70lYAolSwdCzfguQZ',
-              object: 'chat.completion.chunk',
-              // 2023-11-20T09:00:00-05:00
-              created: 1700488800,
-              model: 'gpt-4',
-              choices: [
-                {
-                  index: 0,
-                  finish_reason: null,
-                  delta: { role: 'assistant', content }
-                }
-              ]
-            })
-            this.push(`data: ${chunk}\n\n`)
-            i += 1
-          } else {
-            this.push('data: [DONE]\n\n')
-            this.push(null)
-          }
-        }
-      })
+      const outStream =
+        streamData !== 'do random' ? goodStream(streamData, { ...body }) : badStream({ ...body })
       outStream.pipe(res)
     } else {
       res.write(JSON.stringify(body))
       res.end()
+    }
+  })
+}
+
+function goodStream(dataToStream, chunkTemplate) {
+  const parts = dataToStream.split(' ')
+  let i = 0
+  return new Readable({
+    read() {
+      // This is how the data is streamed from openai
+      // The message response only seems to change and mostly
+      // a stream of content changes via the delta key
+      if (i < parts.length) {
+        const content = parts.length - 1 === i ? parts[i] : `${parts[i]} `
+        chunkTemplate.choices[0].delta.content = content
+        const chunk = JSON.stringify(chunkTemplate)
+        this.push(`data: ${chunk}\n\n`)
+        i += 1
+      } else {
+        this.push('data: [DONE]\n\n')
+        this.push(null)
+      }
+    }
+  })
+}
+
+function badStream(chunkTemplate) {
+  let count = 0
+  return new Readable({
+    read(size = 16) {
+      if (count > 100) {
+        // something is up with OpenAI
+        // you shouldn't have to do this. a throw would be enough
+        chunkTemplate.error = 'exceeded count'
+        this.push('data: ' + JSON.stringify(chunkTemplate) + '\n\n')
+        this.push(null)
+        return
+      }
+
+      const data = crypto.randomBytes(size)
+      chunkTemplate.choices[0].delta.content = data.toString('base64')
+      this.push('data: ' + JSON.stringify(chunkTemplate) + '\n\n')
+      count += 1
     }
   })
 }
